feat: update code structure

+++ src/interrupts/idt.h
@@ -5,19 +5,19 @@
  * How every idt entry is defined
 */
 typedef struct __attribute__((packed)) idt_entry {
-	u16 offset_low; // low offset
-	u16 segment_sel; // segment selector
-	u8 always_zero; // always zero bytes, including reserved bytes
-	u8 flags; // flags
-	u16 offset_high; // high offset
+	uint16_t offset_low; // low offset
+	uint16_t segment_sel; // segment selector
+	uint8_t always_zero; // always zero bytes, including reserved bytes
+	uint8_t flags; // flags
+	uint16_t offset_high; // high offset
 } idt_entry;
 
 /**
  * The IDT table itself
 */
 typedef struct __attribute__((packed)) idt_table {
-	u16 limit; // how many entries this table has?
-	u32 base; // address of the first entry
+	uint32_t limit; // how many entries this table has?
+	uint32_t base; // address of the first entry
 } idt_table;
 
 // array of idt entries
@@ -27,7 +27,7 @@ extern idt_entry idt_entries[NUM_IDT_ENTRIES];
 // idt table itself
 extern idt_table _idt_table;
 
-void create_idt_entry(int index, u32 handler);
+void create_idt_entry(int index, uint32_t handler);
 
 // array of isr handler functions (declared in asm)
 extern void isr_handler_0();
